Separate dimension nodes
