+++ backend/src/nodes/nodes/image/dimension/resize_resolution
@@ -1,0 +1,52 @@
+from __future__ import annotations
+
+import numpy as np
+from sanic.log import logger
+
+from ....categories import ImageDimensionCategory
+from ....node_base import NodeBase
+from ....node_factory import NodeFactory
+from ....properties.inputs import ImageInput, NumberInput, InterpolationInput
+from ....properties.outputs import ImageOutput
+from ....properties import expression
+from ....utils.pil_utils import resize
+
+
+@NodeFactory.register("chainner:image:resize_resolution")
+class ImResizeToResolutionNode(NodeBase):
+    def __init__(self):
+        super().__init__()
+        self.description = (
+            "Resize an image to an exact resolution. "
+            "Auto uses box for downsampling and lanczos for upsampling."
+        )
+        self.inputs = [
+            ImageInput(),
+            NumberInput("Width", minimum=1, default=1, unit="px"),
+            NumberInput("Height", minimum=1, default=1, unit="px"),
+            InterpolationInput(),
+        ]
+        self.outputs = [
+            ImageOutput(
+                image_type=expression.Image(
+                    width="Input1",
+                    height="Input2",
+                    channels="Input0.channels",
+                )
+            )
+        ]
+        self.category = ImageDimensionCategory
+        self.name = "Resize (Resolution)"
+        self.icon = "MdOutlinePhotoSizeSelectLarge"
+        self.sub = "Resize"
+
+    def run(
+        self, img: np.ndarray, width: int, height: int, interpolation: int
+    ) -> np.ndarray:
+        """Takes an image and resizes it"""
+
+        logger.info(f"Resizing image to {width}x{height} via {interpolation}")
+
+        out_dims = (width, height)
+
+        return resize(img, out_dims, interpolation)
